Update SLSTR L1b reader to use core utilities, in preparation for adding L2 support.
